# Disable TUV diagnostics unless debugging