@davegill 2020-07-09
@zxdawn 
Xin,
Thank you for getting this PR ready so quickly.

Would you please explain WHY this change has been made. Does this impact all of WRF Chem? Is it only for certain options? Are the files really large? For the uninitiated, also explain what TUV is. This only needs to be a few sentences. Then take that text and use it in the release notes.