@zxdawn 2020-07-09
@davegill OK. Added the `RELEASE NOTE` above.