@davegill 2020-07-09
@zxdawn 
Xin,
1. Would you post the jenkins test email in these comments.
2. You mention that the processing is faster. Can you quantify that with before vs after timings?