@zxdawn 2020-07-09
1. jenkins test email 

```
Please find result of the WRF regression test cases in the attachment. This build is for Commit ID: 3eac9fe5c4680a704e5172180b4c0f81a0a052a9, requested by: zxdawn for PR: https://github.com/wrf-model/WRF/pull/1243. For any query please send e-mail to David Gill.

    Test Type              | Expected  | Received |  Failed
    = = = = = = = = = = = = = = = = = = = = = = = =  = = = =
    Number of Tests        : 19           18
    Number of Builds       : 48           46
    Number of Simulations  : 166           164        0
    Number of Comparisons  : 105           104        0

    Failed Simulations are: 
    None
    Which comparisons are not bit-for-bit: 
    None
```

2. Time used for TUV init

I just found that if I use 24 cores, the time just cost < 1 second.
But, when I switch to 120 cores, the time costs ~10 minutes when I check the log of `rsl.error.0000`.
BTW, the outputs of `start_datetime` and `end_datetime` defined below are same ...
```
      IF ( 100 .LE. debug_level ) THEN 
        call date_and_time(VALUES=values)
        write(emsg,*)'start_datetime',values
        call wrf_message(trim(emsg))
        call wrf_message('Xin: call diagnostics')
        call diagnostics
        write(emsg,*)'end_datetime',values
        call wrf_message(trim(emsg))
      ENDIF
```