@davegill 2020-07-09
@zxdawn 
Xin,
When you use 24 cores, the job takes < 1 s. When you use 120 cores, the job takes 10 minutes.

I have a couple of questions about this paraphrased statement:
1. Is this the time of the entire job, or the time of the diagnostics call?
2. Is there an improvement in timing when using the new code that bypasses the diagnostics?

Would you fill in the time (s) in a table such as this for the commit message:

|               | Before Commit | After Commit |
| --------| ---------------| -------------  |
| 24 cores |                         |                        |
|120 cores |                        |                        |