@zxdawn 2020-07-15
@davegill Sorry for the late reply. I was preparing some field observations these days.

The time is the time of the diagnostics call.

Here's the table if `debug_level = 0`:

|               | Before Commit | After Commit |
| --------| ---------------| -------------  |
| 24 cores |             1s            |            skip            |
|120 cores |             10 min           |           skip             |

If I change `debug_level` to 100, this is the result:

|               | Before Commit | After Commit |
| --------| ---------------| -------------  |
| 24 cores |             1s            |            1s            |
|120 cores |             10 min           |           10 min             |

I guess there's still some space of improvement for writing the TUV.diags file?
I'm not sure what's the exact problem causing slow writing when using more cores.
I have tested the default I/O setting:
```
 &namelist_quilt
 nio_tasks_per_group = 0,
 nio_groups = 1,
 /
```

The cost time is as same as my personal setting:

```
 &namelist_quilt
 nio_tasks_per_group = 5,
 nio_groups = 2,
 /
```

So, that's not caused by quilting, right?