@davegill 2020-07-15
@zxdawn 
> I have tested the default I/O setting:
The cost time is as same as my personal setting:
So, that's not caused by quilting, right?

Xin, 
I agree. This does not appear to be a quilting issue. The focus of this PR was to reduce the non-debug time. Mission accomplished.