@jordanschnell 2020-07-15
@davegill @zxdawn - Maybe @stacywalters can help provide some insight into the slow write speed, but I am good with the PR as is as well.