@stacywalters 2020-07-15
I did not realize the diagnostic caused performance problems.  The
diagnostic
really was intended for development only.  In all honesty it should have
been
removed from the repository code.

Stacy

On Wed, Jul 15, 2020 at 11:09 AM jordanschnell <notifications@github.com>
wrote:

> @davegill <https://github.com/davegill> @zxdawn
> <https://github.com/zxdawn> - Maybe @stacywalters
> <https://github.com/stacywalters> can help provide some insight into the
> slow write speed, but I am good with the PR as is as well.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/wrf-model/WRF/pull/1243#issuecomment-658820768>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACA4BERT2TIRIVOAKMSBNDTR3XBCZANCNFSM4OUF2C4Q>
> .
>